License update (#6)
* LICENSEed update for better format

Edit of LICENSEed to see how it looks when rendered.

* Test of space on block rendering

* Update LICENSE-ed.md

* Update LICENSE-ed.md

* replaced md licenses with txt versions

GitHub markdown didn't support multiple kinds of numbering of bullets beyond two levels.  But the CC BY 4.0 licenses had four levels of such numbering, each with unique kinds of numerals used (e.g.  a, 1, A, i), so had to revert to straight text forms to match the references to subsections within the license text.

* Minor cleanup of licence.txt indents